feat: bump to kui 12.0.0
This also bumps the node types, and includes a small compilation fix around that.

+++ plugins/plugin-codeflare/src/tray/watchers/profile/status.ts
@@ -66,10 +66,12 @@ export default class ProfileStatusWatcher {
 
     // make sure to kill that watcher subprocess when we exit
     process.on("exit", () => {
-      try {
-        process.kill(-job.pid) // kill the process group e.g. for pipes
-      } catch (err) {
-        Debug("codeflare")("error killing process group " + -job.pid, err)
+      if (job.pid) {
+        try {
+          process.kill(-job.pid) // kill the process group e.g. for pipes
+        } catch (err) {
+          Debug("codeflare")("error killing process group " + -job.pid, err)
+        }
       }
 
       try {
